Leaves

+++ 12-binary_tree_leaves.c
@@ -13,7 +13,7 @@ size_t binary_tree_leaves(const binary_tree_t *tree)
 
 	if (tree)
 	{
-		leaves += binary_tree_is_leaf(tree);
+		leaves += (size_t) binary_tree_is_leaf(tree);
 		leaves += binary_tree_leaves(tree->left);
 		leaves += binary_tree_leaves(tree->right);
 	}
